move codec within kernel
